requirements.txt
django==4.1.1
djangorestframework==3.14.0
gunicorn==20.1.0
requests==2.28.1
urllib3==1.26.12
colorhash==1.2.1
pyarr==3.1.3